new fast compiler prototype
