+++ package.json
@@ -1,93 +1,105 @@
 {
-    "name": "@seagull/framework",
-    "version": "0.1.0",
-    "description": "Complete core and utils for developing seagull apps",
-    "browser": "dist/src/core/browser.js",
-    "main": "dist/src/index.js",
-    "types": "dist/src/index.d.js",
-    "scripts": {
-        "build": "rm -rf dist && tsc",
-        "test:functional": "NODE_ENV=test mocha --opts test/functional/mocha.opts",
-        "test:functional:watch": "npm run test:functional -- --watch --reporter min",
-        "test:unit": "NODE_ENV=test mocha --opts test/unit/mocha.opts",
-        "test:unit:watch": "npm run test:unit -- --watch --reporter min",
-        "test": "NODE_ENV=test nyc mocha --opts test/mocha.opts",
-        "test:watch": "npm run test -- --watch --reporter min",
-        "docs": "rm -rf docs && typedoc --mode modules --excludeExternals --excludePrivate --excludeNotExported --out docs src"
-    },
-    "repository": {
-        "type": "git",
-        "url": "git+https://github.com/seagull-js/framework.git"
-    },
-    "author": "Maximilian Stroh <hisako1337@gmail.com>",
-    "license": "LGPL-3.0",
-    "devDependencies": {
-        "@types/chai": "^4.1.2",
-        "@types/history": "^4.6.2",
-        "@types/lodash": "^4.14.103",
-        "@types/mocha": "^2.2.48",
-        "@types/node": "^6.0.100",
-        "aws-sdk": "^2.209.0",
-        "aws-sdk-mock": "^1.7.0",
-        "chai": "^4.1.2",
-        "mobx": "^3.3.1",
-        "mobx-react": "^4.3.3",
-        "mobx-react-router": "^4.0.1",
-        "mocha": "^5.0.1",
-        "mocha-typescript": "^1.1.12",
-        "mock-fs": "^4.4.2",
-        "mock-require": "^3.0.2",
-        "nyc": "^11.6.0",
-        "source-map-support": "^0.5.3",
-        "ts-node": "^4.1.0",
-        "tsconfig-paths": "^3.1.1",
-        "tslint": "^5.9.1",
-        "tslint-config-prettier": "^1.8.0",
-        "typedoc": "^0.11.1",
-        "typedoc-plugin-external-module-name": "^1.1.1",
-        "typescript": "^2.7.2"
-    },
-    "dependencies": {
-        "@types/react": "^16.1.0",
-        "@types/react-dom": "^16.0.4",
-        "aws-lambda": "^0.1.2",
-        "browserify": "^16.1.1",
-        "browserify-incremental": "^3.1.1",
-        "clime": "^0.5.9",
-        "cloudformation-declarations": "^0.1.2",
-        "csstips": "^0.2.3",
-        "dashify": "^1.0.0",
-        "decache": "^4.4.0",
-        "lodash": "^4.17.5",
-        "node-dir": "^0.1.17",
-        "package-json": "^4.0.1",
-        "prettier": "^1.10.2",
-        "react": "^16.3.0",
-        "react-dom": "^16.3.0",
-        "react-helmet": "^5.2.0",
-        "react-router": "^4.2.0",
-        "react-router-config": "^1.0.0-beta.4",
-        "react-router-dom": "^4.2.2",
-        "reflect-metadata": "^0.1.12",
-        "require-from-string": "^2.0.1",
-        "shelljs": "^0.8.1",
-        "shortid": "^2.2.8",
-        "stoppable": "^1.0.5",
-        "stream-to-string": "^1.1.0",
-        "ts-simple-ast": "^8.1.0",
-        "typescript-memoize": "^1.0.0-alpha.3",
-        "typestyle": "^1.7.2",
-        "uglify-es": "^3.3.9",
-        "yamljs": "^0.3.0"
-    },
-    "peerDependencies": {
-        "aws-sdk": "^2.209.0",
-        "typescript": "^2.7.2"
-    },
-    "nyc": {
-        "extension": [".ts", ".tsx"],
-        "exclude": ["**/*.d.ts", "dist"],
-        "reporter": ["text-summary", "html"],
-        "all": true
-    }
+  "name": "@seagull/framework",
+  "version": "0.1.0",
+  "description": "Complete core and utils for developing seagull apps",
+  "browser": "dist/src/core/browser.js",
+  "main": "dist/src/index.js",
+  "types": "dist/src/index.d.js",
+  "scripts": {
+    "build": "rm -rf dist && tsc",
+    "test:functional": "NODE_ENV=test mocha --opts test/functional/mocha.opts",
+    "test:functional:watch": "npm run test:functional -- --watch --reporter min",
+    "test:unit": "NODE_ENV=test mocha --opts test/unit/mocha.opts",
+    "test:unit:watch": "npm run test:unit -- --watch --reporter min",
+    "test": "NODE_ENV=test nyc mocha --opts test/mocha.opts",
+    "test:watch": "npm run test -- --watch --reporter min",
+    "docs": "rm -rf docs && typedoc --mode modules --excludeExternals --excludePrivate --excludeNotExported --out docs src"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/seagull-js/framework.git"
+  },
+  "author": "Maximilian Stroh <hisako1337@gmail.com>",
+  "license": "LGPL-3.0",
+  "devDependencies": {
+    "@types/chai": "^4.1.2",
+    "@types/chokidar": "^1.7.5",
+    "@types/history": "^4.6.2",
+    "@types/lodash": "^4.14.103",
+    "@types/mocha": "^2.2.48",
+    "@types/node": "^6.0.100",
+    "aws-sdk": "^2.209.0",
+    "aws-sdk-mock": "^1.7.0",
+    "chai": "^4.1.2",
+    "mobx": "^3.3.1",
+    "mobx-react": "^4.3.3",
+    "mobx-react-router": "^4.0.1",
+    "mocha": "^5.0.1",
+    "mocha-typescript": "^1.1.12",
+    "mock-fs": "^4.4.2",
+    "mock-require": "^3.0.2",
+    "nyc": "^11.6.0",
+    "source-map-support": "^0.5.3",
+    "ts-node": "^4.1.0",
+    "tsconfig-paths": "^3.1.1",
+    "tslint": "^5.9.1",
+    "tslint-config-prettier": "^1.8.0",
+    "typedoc": "^0.11.1",
+    "typedoc-plugin-external-module-name": "^1.1.1",
+    "typescript": "^2.7.2"
+  },
+  "dependencies": {
+    "@types/react": "^16.1.0",
+    "@types/react-dom": "^16.0.4",
+    "aws-lambda": "^0.1.2",
+    "browserify": "^16.1.1",
+    "browserify-incremental": "^3.1.1",
+    "chokidar": "^2.0.3",
+    "clime": "^0.5.9",
+    "cloudformation-declarations": "^0.1.2",
+    "copy-node-modules": "^1.0.4",
+    "csstips": "^0.2.3",
+    "dashify": "^1.0.0",
+    "decache": "^4.4.0",
+    "lodash": "^4.17.5",
+    "node-dir": "^0.1.17",
+    "package-json": "^4.0.1",
+    "prettier": "^1.10.2",
+    "react": "^16.3.0",
+    "react-dom": "^16.3.0",
+    "react-helmet": "^5.2.0",
+    "react-router": "^4.2.0",
+    "react-router-config": "^1.0.0-beta.4",
+    "react-router-dom": "^4.2.2",
+    "reflect-metadata": "^0.1.12",
+    "require-from-string": "^2.0.1",
+    "shelljs": "^0.8.1",
+    "shortid": "^2.2.8",
+    "stoppable": "^1.0.5",
+    "stream-to-string": "^1.1.0",
+    "ts-simple-ast": "^8.1.0",
+    "typescript-memoize": "^1.0.0-alpha.3",
+    "typestyle": "^1.7.2",
+    "uglify-es": "^3.3.9",
+    "yamljs": "^0.3.0"
+  },
+  "peerDependencies": {
+    "aws-sdk": "^2.209.0",
+    "typescript": "^2.7.2"
+  },
+  "nyc": {
+    "extension": [
+      ".ts",
+      ".tsx"
+    ],
+    "exclude": [
+      "**/*.d.ts",
+      "dist"
+    ],
+    "reporter": [
+      "text-summary",
+      "html"
+    ],
+    "all": true
+  }
 }
